refactor(API): reduce to an unique constant
