+++ api/src/quest/domain/models/combined-course-participations/aggregates/CombinedCourseDetails.js
@@ -4,6 +4,7 @@ import {
 } from '../../../../../prescription/shared/domain/constants.js';
 import { config } from '../../../../../shared/config.js';
 import { cryptoService as injectedCryptoService } from '../../../../../shared/domain/services/crypto-service.js';
+import { COMBINED_COURSE_ITEM_TYPES } from '../../../constants.js';
 import { CombinedCourse } from '../../combined-courses/entities/CombinedCourse.js';
 import { DataForQuest } from '../../quests/aggregates/DataForQuest.js';
 import { Eligibility } from '../../quests/aggregates/Eligibility.js';
@@ -11,7 +12,6 @@ import { Quest, REQUIREMENT_TYPES } from '../../quests/entities/Quest.js';
 import { TYPES } from '../../quests/value-objects/Requirement.js';
 import {
   CampaignCombinedCourseItem,
-  COMBINED_COURSE_ITEM_TYPES,
   ModuleCombinedCourseItem,
   TrainingCombinedCourseItem,
 } from '../value-objects/CombinedCourseItem.js';
